allowed changing of the water colour, map info editor refactor
