+++ src/utilities/file_paths.ixx
@@ -16,4 +16,9 @@ namespace paths {
 		return hive_data_folder(map_directory) / "terrain_pathing.json";
 	}
 
+	/// Returns the path to the map info file
+	export fs::path map_info_extras_file(const fs::path& map_directory) {
+		return hive_data_folder(map_directory) / "map_info_extras.json";
+	}
+
 } // namespace paths
